requirements.txt
python-dotenv
pandas
psycopg2-binary
sqlalchemy
xarray
netCDF4
numpy
